Rename expr/cnf.rs to expr.rs
